+++ backend/app/schemas/insights/category_trend_schema.py
@@ -1,10 +1,11 @@
+from decimal import Decimal
 from typing import List, Literal, Optional
 
 from pydantic import BaseModel
 
 
 class CategoryTrendDataPoint(BaseModel):
-    amount: float
+    amount: Decimal
     category_name: str
