+++ backend/app/schemas/insights/fund_flow_schema.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import List, Literal, Optional
 
 from pydantic import BaseModel
@@ -11,9 +12,9 @@ class FundFlowNode(BaseModel):
     currency_symbol: str
     account_type: str
     account_subtype: str
-    total_outflow: float
-    total_inflow: float
-    total_volume: float
+    total_outflow: Decimal
+    total_inflow: Decimal
+    total_volume: Decimal
 
 
 class FundFlowLink(BaseModel):
@@ -23,22 +24,22 @@ class FundFlowLink(BaseModel):
     target_account_id: int
     target_account_name: str
     target_ledger_name: str
-    total_amount: float
+    total_amount: Decimal
     transfer_count: int
     is_cross_ledger: bool
-    target_amount: Optional[float] = None
+    target_amount: Optional[Decimal] = None
     source_currency: str
     target_currency: str
 
 
 class FundFlowSummary(BaseModel):
-    total_transfer_volume: float
+    total_transfer_volume: Decimal
     total_transfer_count: int
     cross_ledger_count: int
     unique_corridors: int
     most_active_corridor_source: Optional[str] = None
     most_active_corridor_target: Optional[str] = None
-    most_active_corridor_amount: Optional[float] = None
+    most_active_corridor_amount: Optional[Decimal] = None
 
 
 class FundFlowResponse(BaseModel):
