+++ backend/app/schemas/insights/store_location_schema.py
@@ -1,3 +1,4 @@
+from decimal import Decimal
 from typing import List, Literal
 
 from pydantic import BaseModel
@@ -5,44 +6,44 @@ from pydantic import BaseModel
 
 class StoreExpenseData(BaseModel):
     store: str
-    amount: float
-    percentage: float
+    amount: Decimal
+    percentage: Decimal
 
 
 class LocationExpenseData(BaseModel):
     location: str
-    amount: float
-    percentage: float
+    amount: Decimal
+    percentage: Decimal
 
 
 class ExpenseByStoreResponse(BaseModel):
     store_data: List[StoreExpenseData]
-    total_expense: float
+    total_expense: Decimal
     period_type: Literal["all_time", "last_12_months", "this_month"]
 
 
 class ExpenseByLocationResponse(BaseModel):
     location_data: List[LocationExpenseData]
-    total_expense: float
+    total_expense: Decimal
     period_type: Literal["all_time", "last_12_months", "this_month"]
 
 
 class CategoryExpenseData(BaseModel):
     category_id: int
     category: str
-    amount: float
-    percentage: float
+    amount: Decimal
+    percentage: Decimal
 
 
 class StoreCategoryBreakdownResponse(BaseModel):
     store: str
     category_data: List[CategoryExpenseData]
-    total_expense: float
+    total_expense: Decimal
     period_type: str
 
 
 class LocationCategoryBreakdownResponse(BaseModel):
     location: str
     category_data: List[CategoryExpenseData]
-    total_expense: float
+    total_expense: Decimal
     period_type: str
